add build dependencies to README.md
This commit adds the build dependencies
for Debian/Ubuntu and RedHat/Fedora systems
to the README - since running just `cargo build`
does not work on its own.

It does not add any build dependency information
for MacOS or Windows systems (or other linux distros).
In particular, a Mac user may should try to do a fresh
clone and `cargo build` and add any missing dependencies
to the README.